requirements.txt
constriction
numpy
pandas
pillow
pytorch-msssim
PyYAML
tqdm
vector-quantize-pytorch
jaxtyping
Ninja
# Add opencv for video loading
opencv-python
einops
scikit-learn
matplotlib
# gsplat @ git+https://github.com/hbb1/gsplat@main
